Update test suite, TravisCI integration
(Work in progress)

Prior to this PR, we used a testing setup introduced in #50 by @SleeplessByte. The process required checking out a separate repository and configuring it to work with PHPUnit. This PR introduces a workflow described in [this article]() that automates this a bit. To begin using the testing suite, run `./bin/install-wp-tests.sh` with the following arguments:

```bash
$ ./bin/install-wp-tests.sh <db-name> <db-user> <db-pass> [db-host] [wp-version]
```

For example, if you wanted to install it to a test database named `wordpress_test` as the `root` user with no password on localhost, you would run:

```bash
$ ./bin/install-wp-tests.sh wordpress_test root '' 127.0.0.1
```

With this in place, running `phpunit` from the command line (installed via PEAR, PHAR or Composer) will run correctly because it looks for the version of WordPress installed to the `/tmp` directory.

---

Current test failures that need to be addressed:

```
Stephens-Air:schema-creator (feature/fixup-test-suite) stephen$ phpunit
Installing...
Running as single site... To run multisite, use -c tests/phpunit/multisite.xml
Not running ajax tests. To execute these, use --group ajax.
Not running ms-files tests. To execute these, use --group ms-files.
Not running external-http tests. To execute these, use --group external-http.
PHPUnit 4.5.0 by Sebastian Bergmann and contributors.

Configuration read from /Users/stephen/Sites/schema-creator/phpunit.xml

..FF..FFFR.....

Time: 1.6 seconds, Memory: 22.00Mb

There were 5 failures:

1) BasicTest::testMetaBox
metabox is hidden when it should not be
Failed asserting that false is true.

/Users/stephen/Sites/schema-creator/tests/SchemaCreator/BasicTest.php:82

2) BasicTest::testStoreSettings
default css option is not false but false
Failed asserting that false matches expected 'false'.

/Users/stephen/Sites/schema-creator/tests/SchemaCreator/BasicTest.php:153

3) BasicTest::testAdminScriptsSettingsPage
Admin schema style is not enqueued.
Failed asserting that false is true.

/Users/stephen/Sites/schema-creator/tests/SchemaCreator/BasicTest.php:203

4) BasicTest::testSchemaFooter
footer is not altered when on settings page
Failed asserting that a string is not empty.

/Users/stephen/Sites/schema-creator/tests/SchemaCreator/BasicTest.php:218

5) BasicTest::testBodyClass
Itemscope not inserted by default.
Failed asserting that '' contains "itemscope".

/Users/stephen/Sites/schema-creator/tests/SchemaCreator/BasicTest.php:240
```